+++ GeometryDemo/IQGeometry/IQGeometry+Angle.c
@@ -6,8 +6,9 @@
 //  Copyright (c) 2013 Iftekhar. All rights reserved.
 //
 
-#import "IQGeometry+Angle.h"
-#import "IQGeometry+Distance.h"
+#include "IQGeometry+Angle.h"
+#include "IQGeometry+Distance.h"
+#include <math.h>
 
 CGFloat IQDegreeToRadian(CGFloat angle)
 {
@@ -22,9 +23,8 @@ CGFloat IQRadianToDegree(CGFloat radians)
 CGFloat IQPointGetAngle(CGPoint centerPoint, CGPoint point1, CGPoint point2)
 {
     //Find angle.
-    CGFloat angle = atan2(point2.y-centerPoint.y, point2.x-centerPoint.x) - atan2(point1.y-centerPoint.y, point1.x-centerPoint.x);
+    CGFloat angle = atan2f(point2.y-centerPoint.y, point2.x-centerPoint.x) - atan2f(point1.y-centerPoint.y, point1.x-centerPoint.x);
 
     //If angle is less than 0. then adding 360 degree to it.
     return (angle>0 ? angle : angle+M_PI*2);
 }
-
